rent fixed

+++ src/main/java/com/andrew/rental/service/ActiveRentService.java
@@ -1,7 +1,9 @@
 package com.andrew.rental.service;
 
 import com.andrew.rental.domain.ActiveRent;
+import com.andrew.rental.domain.Car;
 import com.andrew.rental.domain.User;
+import com.andrew.rental.domain.enumeration.CarStatus;
 import com.andrew.rental.repository.ActiveRentRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,6 +13,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
+import java.time.Duration;
 import java.util.Optional;
 
 /**
@@ -24,9 +28,19 @@ public class ActiveRentService {
 
     private final ActiveRentRepository activeRentRepository;
 
-    public ActiveRentService(ActiveRentRepository activeRentRepository) {
+    private final PaymentService paymentService;
+
+    private final CarService carService;
+
+    public ActiveRentService(ActiveRentRepository activeRentRepository,
+                             PaymentService paymentService,
+                             CarService carService) {
         this.activeRentRepository = activeRentRepository;
+        this.paymentService = paymentService;
+        this.carService = carService;
     }
+
+    private final double tax = 0.3;
 
     /**
      * Save a activeRent.
@@ -34,7 +48,34 @@ public class ActiveRentService {
      * @param activeRent the entity to save.
      * @return the persisted entity.
      */
-    public ActiveRent save(ActiveRent activeRent) {
+    public ActiveRent rent(ActiveRent activeRent) throws IllegalAccessException {
+        User client = activeRent.getClient();
+        Car car = activeRent.getCar();
+
+        if (car.getStatus() != CarStatus.AVAILABLE) {
+            throw new IllegalAccessException("The car is not available");
+        }
+
+        User owner = car.getOwner();
+
+        Duration duration = activeRent.getDuration();
+
+        BigDecimal amount = car.getPricePerHour().
+            multiply(BigDecimal.valueOf(duration.toHours() * (1 - tax)));
+
+        if (client.getIban()==null || owner.getIban()==null) {
+            throw new IllegalAccessException("No bank card");
+
+        }
+        boolean success = paymentService.
+            transaction(client.getIban(), owner.getIban(), amount);
+
+        if (!success) {
+            throw new IllegalAccessException("Try again");
+        }
+
+        carService.setStatusById(car.getId(), CarStatus.RENTED);
+
         log.debug("Request to save ActiveRent : {}", activeRent);
         return activeRentRepository.save(activeRent);
     }
@@ -75,8 +116,16 @@ public class ActiveRentService {
      *
      * @param id the id of the entity.
      */
-    public void delete(Long id) {
+    public void closeRent(Long id) throws IllegalAccessException {
         log.debug("Request to delete ActiveRent : {}", id);
+        Optional<ActiveRent> rent = activeRentRepository.findById(id);
+        if (!rent.isPresent()) {
+            throw new IllegalAccessException("Bad request");
+        }
+
+        carService.setStatusById(rent.get().getCar().getId(),
+            CarStatus.AVAILABLE);
+
         activeRentRepository.deleteById(id);
     }
 }
